cleanup leftovers items
Cleanup leftover items from initial code.  Tried doing templates a
couple different ways.  There is a default template that gets embedded
but you use that as a baseline and place the
AggregateRootCoreTemplate.t4 file in the same folder as your Dizzle
files and it will use that one for code gen.